% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuarsadd-package.R
\docType{package}
\name{wuarsadd}
\alias{wuarsadd}
\alias{wuarsadd-package}
\alias{wuarsadd-package}
\title{wuarsadd}
\description{
Wilson's Useful and Awesome RStudio ADDins
}
\author{
\strong{Maintainer}: Wilson Freitas \email{wilson.freitas@gmail.com}

}
\keyword{internal}
